시작 위치 application으로 변경
