--- src/test/java/MainTest.java
@@ -1,34 +1,0 @@
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class MainTest {
-
-    @Test
-    void initList() {
-    }
-
-    @Test
-    void makeAttack() {
-    }
-
-    @Test
-    void makeAns() {
-    }
-
-    @Test
-    void compareAnswer() {
-    }
-
-    @Test
-    void makePredict() {
-    }
-
-    @Test
-    void printCompareResult() {
-    }
-
-    @Test
-    void main() {
-    }
-}
